aider: Added a new function "search_database" to the `functions` list and adjusted the parameters of `_construct_prompt`.
# Aider chat conversation:

USER: We need to add a new function in the `functions` list with the name of "search_database" the description will need to be something like: "You have access to a rich Knowledge Database however in order to be used you must ask a list of questions that will be used to search the database please reply with a list of questions that would aid in answer the user's quesion" remember to adjust the paremeters of `_construct_prompt`

ASSISTANT: Sure, we can add a new function to the `functions` list and adjust the parameters of the `_construct_prompt` method. Here's how you can do it:

```python
backend/llm/openai_functions.py
<<<<<<< ORIGINAL
    def generate_answer(self, question: str) -> ChatHistory:
        ...
        functions = [
            {
                "name": "get_history_and_context",
                "description": "Get the chat history between you and the user and also get the relevant documents to answer the question. Always use that unless a very simple question is asked that a 5 years old could answer.",
                "parameters": {"type": "object", "properties": {}},
            },
        ]
        ...
        response = self._get_model_response(
            messages=self._construct_prompt(question), functions=functions
        )
        ...
=======
    def generate_answer(self, question: str) -> ChatHistory:
        ...
        functions = [
            {
                "name": "get_history_and_context",
                "description": "Get the chat history between you and the user and also get the relevant documents to answer the question. Always use that unless a very simple question is asked that a 5 years old could answer.",
                "parameters": {"type": "object", "properties": {}},
            },
            {
                "name": "search_database",
                "description": "You have access to a rich Knowledge Database however in order to be used you must ask a list of questions that will be used to search the database please reply with a list of questions that would aid in answer the user's quesion",
                "parameters": {"type": "object", "properties": {}},
            },
        ]
        ...
        response = self._get_model_response(
            messages=self._construct_prompt(question, functions=functions), functions=functions
        )
        ...
>>>>>>> UPDATED
```

In this change, we added a new function to the `functions` list with the name "search_database" and the specified description. We also adjusted the parameters of the `_construct_prompt` method to accept the `functions` list.