aider: Update `search_database` function in `openai_functions.py` to clarify that the model should respond with a list of questions and add a `questions` parameter to the function.
# Aider chat conversation:

USER: what information does `formatted_response` hold I want to be able to extract a list of questions from the response

ASSISTANT: The `formatted_response` is an instance of the `OpenAiAnswer` class. It contains two attributes:

1. `content`: This is the actual text of the response from the model.
2. `function_call`: This is an instance of the `FunctionCall` class, which contains the name of the function that the model called and the arguments it passed to that function.

The `FunctionCall` class has two attributes:

1. `name`: The name of the function that the model called.
2. `arguments`: The arguments that the model passed to the function. This is a dictionary where the keys are the argument names and the values are the argument values.

If the model is expected to return a list of questions in the `arguments` of a function call, you can extract them like this:

```python
questions = formatted_response.function_call.arguments['questions']
```

This assumes that the model is returning the questions in an argument named 'questions'. If the argument has a different name, you should replace 'questions' with the correct argument name.

USER: I like that approach but we also need to modify `functions` so that the model knows to respond in the correct manner

ASSISTANT: To modify the `functions` so that the model knows to respond with a list of questions when it calls `search_database`, we need to update the description and parameters of the `search_database` function in the `generate_answer` method of the `OpenAIFunctionsBrainPicking` class in the `backend/llm/openai_functions.py` file.

Here is the change:

```python
backend/llm/openai_functions.py
<<<<<<< ORIGINAL
        functions = [
            {
                "name": "get_history_and_context",
                "description": "Get the chat history between you and the user and also get the relevant documents to answer the question. Always use that unless a very simple question is asked that a 5 years old could answer.",
                "parameters": {"type": "object", "properties": {}},
            },
            {
                "name": "search_database",
                "description": "You have access to a rich Knowledge Database however in order to be used you must ask a list of questions that will be used to search the database please reply with a list of questions that would aid in answer the user's quesion",
                "parameters": {"type": "object", "properties": {}},
            },
        ]
=======
        functions = [
            {
                "name": "get_history_and_context",
                "description": "Get the chat history between you and the user and also get the relevant documents to answer the question. Always use that unless a very simple question is asked that a 5 years old could answer.",
                "parameters": {"type": "object", "properties": {}},
            },
            {
                "name": "search_database",
                "description": "You have access to a rich Knowledge Database. To use it, please reply with a list of questions that would aid in answering the user's question.",
                "parameters": {"type": "object", "properties": {"questions": {"type": "array", "items": {"type": "string"}}}},
            },
        ]
>>>>>>> UPDATED
```

This change updates the description of the `search_database` function to clarify that the model should reply with a list of questions. It also adds a `questions` parameter to the function, which is an array of strings. This tells the model to return the questions in this parameter.